requirements.txt
streamlit
pandas
qrcode
cryptography
Pillow
pyzbar
